Adding new logic
This update makes sure a new random quote is shown each time the button is clicked on on the setInterval time.  Like choose names out of a hat, its random and there are less choices each time to pick from.  And when the number of items goes down to zero and a user clicks the button again, then the array is populated again and all the quotes return.